+++ buildy-wp-gui/src/components/TextModule.vue
@@ -2,6 +2,7 @@
   <settings-modal>
     <title-editor label="Title" path="content.title" />
     <component :is="editorType" path="content.body"></component>
+    <custom-fields />
     <div slot="options" class="text-custom-options">
       <div class="flex">
         <toggle-switch