+++ src/js/main.js
@@ -1,5 +1,8 @@
 import accordions from './components/accordion'
 import { initSliders } from './components/slider'
+import baguetteBox from 'baguettebox.js';
+
+baguetteBox.run('.bmcb-gallery:not(.bmcb-slider) .bmcb-gallery__items')
 
 accordions();